requirements.txt
black==23.3.0
click==8.1.3
isort==5.12.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
tomli==2.0.1
